Fix: extract shouldCheckForUpdate to avoid side-effect-on-import
Importing main.ts to test shouldCheckForUpdate triggered main().catch() at the top level, running the full CLI inside vitest and crashing with process.exit(1). Extracted the function to src/util/update-check.ts which has no side effects. Tests now import safely.

Lore-id: b204b8d0
Constraint: main.ts must never be imported by tests — it has top-level side effects
Constraint: Extracted module must be a pure function with no imports from services
Rejected: Entry point guard (if isEntryPoint) | couples main.ts to import.meta.url checks
Rejected: Environment guard (if !VITEST) | couples production code to test runner
Rejected: Keeping function in main.ts with a different export pattern | root cause is the top-level main() call, not the export
Tested: All 10 shouldCheckForUpdate tests pass without process.exit crash
Tested: Full test suite 417 tests pass
Confidence: high
Scope-risk: narrow
Reversibility: clean
